requirements.txt
absl-py==2.1.0
antlr4-python3-runtime==4.9.3
astunparse==1.6.3
black==24.3.0
blinker==1.7.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
cloudpickle==3.0.0
contourpy==1.2.0
cycler==0.12.1
detectron2 @ git+https://github.com/facebookresearch/detectron2.git@afe9eb920646102f7e6bf0cd2115841cea2aca13
filelock==3.9.0
Flask==3.0.2
flatbuffers==24.3.25
fonttools==4.50.0
fsspec==2023.4.0
fvcore==0.1.5.post20221221
gast==0.5.4
google-pasta==0.2.0
grpcio==1.62.1
h5py==3.10.0
hydra-core==1.3.2
idna==3.6
iopath==0.1.9
itsdangerous==2.1.2
Jinja2==3.1.2
keras==3.1.1
kiwisolver==1.4.5
libclang==18.1.1
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.8.3
mdurl==0.1.2
ml-dtypes==0.3.2
mpmath==1.3.0
mypy-extensions==1.0.0
namex==0.0.7
networkx==3.2.1
numpy==1.26.3
omegaconf==2.3.0
opt-einsum==3.3.0
optree==0.11.0
packaging==24.0
pathspec==0.12.1
pillow==10.2.0
platformdirs==4.2.0
portalocker==2.8.2
protobuf==4.25.3
pycocotools==2.0.7
Pygments==2.17.2
pyparsing==3.1.2
python-dateutil==2.9.0.post0
PyYAML==6.0.1
requests==2.31.0
rich==13.7.1
six==1.16.0
sympy==1.12
tabulate==0.9.0
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorflow==2.16.1
tensorflow-io-gcs-filesystem==0.36.0
termcolor==2.4.0
torch==2.2.2+cpu
torchaudio==2.2.2+cpu
torchvision==0.17.2+cpu
tqdm==4.66.2
typing_extensions==4.8.0
urllib3==2.2.1
Werkzeug==3.0.1
wrapt==1.16.0
yacs==0.1.8
